feat: Added BlurImageLoader component
